% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{check_raster_overlaps}
\alias{check_raster_overlaps}
\title{Check Raster Overlaps}
\usage{
check_raster_overlaps(fdir_filtered, epsg = NULL, action = "stop")
}
\arguments{
\item{fdir_filtered}{An \code{fdir} dataframe / sf object}

\item{epsg}{Only necessary if \code{fdir} is not an sf object}
}
\description{
Takes an \code{fdir} object \strong{with} geometries and checks if any of the
features overlap. This function is only in \code{get_raster()} after filtering
and before merging.
}
