% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_colours.R
\name{colortable_greyscale}
\alias{colortable_greyscale}
\title{Colortable of singleband Raster into Greyscale}
\usage{
colortable_greyscale(colortable, method = "ysrgb")
}
\arguments{
\item{method}{The method to use when turning RGB to greyscale.}

\item{raster_in}{An object of class RasterBrick.}
}
\description{
\code{rgb_raster2singleband} turns a threeband RGB Raster (\code{\link{brick}})
into a singleband GREYSCALE \code{\link{raster}} with a \code{\link{colortable}}
}
