% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdir_helper.R
\name{credits}
\alias{credits}
\title{Return String with credits}
\usage{
credits(who = "swisstopo")
}
\arguments{
\item{who}{Character string specifying whom to credit. Currently only "swisstopo" is implemented.}
}
\description{
Perticularly swisstopo wants to be credited with the special Copyright character (\u00A9)
Since I need to credit the everytime I make a map, i've prepared a function returing
that exact text.
}
