% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdir_other.R
\name{fdir_export}
\alias{fdir_export}
\title{Export an existing fdir to an Rda-File}
\usage{
fdir_export(path)
}
\description{
Export an existing fdir to an Rda-File
}
\details{
The function enables exporting an existing "File Directory"
(\code{fdir}) to an ".Rda" File after running \code{fdir_init}.
This avoids having to rescan all the files with \code{fdir_init}.
Handle with care, changes in the source files are not registered with
this method. Use \code{fdir_init} if unsure.
}
