% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_helper.R
\name{fdir_filter}
\alias{fdir_filter}
\title{Filter fdir to find the required raster files}
\usage{
fdir_filter(fdir, epsg, scale_level, xmin, xmax, ymin, ymax, year = NULL,
  name = "")
}
\arguments{
\item{fdir}{A dataframe or \pkg{sf} object from \code{fdir_init}}

\item{xmin, xmax, ymin, ymax}{Integers specyfing the extent of the desired raster}

\item{year}{Optional integer specyfing the year from which data is needed}

\item{scale}{level integer specyfing the scale level of the raster}
}
\description{
Filters an \code{fdir} object to find the required raster files
Only used in \code{get_raster()}. Outsorced it to a function in order to
modularize \code{get_raster()} and make it easier to debug.
}
