% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fdir_import}
\alias{fdir_import}
\title{Import an existing fdir}
\usage{
fdir_import(path)
}
\arguments{
\item{name}{Path to an .Rda File}
}
\description{
Import an existing fdir
}
\details{
The function enables importing an existing "File Directory"
(\code{fdir}) from an ".Rda" File after running \code{fdir_init}
and \code{fdir_export}.
This avoids having to rescan all the files with \code{fdir_init}.
Handle with care, changes in the source files are not registered with
this method. Use \code{fdir_init} if unsure.
}
