% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{fdir_init}
\alias{fdir_init}
\title{Initialize File Directory}
\usage{
fdir_init(rootdir, maxfiles = Inf, add_geometry = T, filter = c("PK",
  "SMR", "LK", "TA"))
}
\arguments{
\item{rootdir}{Character string specifying the directory where the folders are stored OR
character string pointing to an .Rda file from a previous fdir_init() run OR an fdir variable.}

\item{maxfiles}{Integer limiting the number of files to be scanned. For testing purposes only.}

\item{add_geometry}{Should the bounding box of each file be added as a geometry to \code{fdir}?}

\item{filter}{Names of the folders to look for. Only folders containing the character strings
specified here are included in the search.}
}
\description{
Scans all folders in the root dirctory corresponding to the folders given by
\code{folders = }. OR: Makes an existing fdir object (either a variable or
an .Rda File) the current fdir file by writing it to the Environment
'swissrastermapEnv'. If stored as an Rda File, the associated variable name
must be \code{fdir}
}
\details{
This command creates a "File Directory" in the package environment by scanning
all folders specified by \code{folders} within the \code{rootdir} and analyzing the content.
All files ending with "tif" are checked for extent, number of layers and resolution.
All the mentioned attributes of each raster file, along with the file path and the extent as a
geometry, are stoerd in the variable \code{fdir} of the package environment.
Parent folders containing the rasterdata require a specific structure:
\code{TYPE_Scale_EPSG_index}
\describe{
  \item{\code{TYPE}}{corresponds to the maptype. IN Switzerland this is typically \code{SMR}, \code{PK}, \code{LK} or \code{TA}. \strong{Must} be followed by a \code{_}}
  \item{\code{Scale}}{defines the map scale as x in 1:1'000x. \strong{Must} be followed by a \code{_}}
  \item{\code{EPSG}}{specifies the CRS of the containing raster data \strong{Must} be followed by a \code{_} \strong{only} if \code{name} is specified}
  \item{\code{index}}{Needed when multiple folders with same maptype, scale and epsg code exist, but different naming patterns.}
}
}
