% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gemeinden_top_centroid}
\alias{gemeinden_top_centroid}
\title{Gemeinden Top Centroid}
\format{An \code{sf} object containing point data
\describe{
  \item{NAME}{Name of the muncipality in its main language}
  \item{EINWOHNERZ}{Number of inhabitants}
  \item{geometry}{sfc_POINT data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
gemeinden_top_centroid
}
\description{
Centroids of the 10 largest muncipalities ("Gemeinden") in Switzerland.
This dataset is based on \code{\link{gemeinden_top_centroid}}
}
\keyword{datasets}
