% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{geom_from_boundary}
\alias{geom_from_boundary}
\title{Make Geometry from boundary}
\usage{
geom_from_boundary(df, epsg, add = T)
}
\arguments{
\item{df}{A dataframe including the columns \code{xmin}, \code{xmax},
\code{ymin}, \code{max}}

\item{epsg}{Intger value of the CRS the data is set to.}

\item{add}{A boolean determining whether the geometries should be added to the
input dataframe (\code{add = TRUE}) or if only the geometries should be returned
(\code{add = FALSE}).}
}
\description{
Takes a dataframe at returns an \pkg{sf} object.
}
\details{
This function takes a dataframe and returns an \pkg{sf} object. The dataframe
must contain the columns, \code{xmin}, \code{xmax}, \code{ymin}, \code{max},
which will be turned into the edges of a rectangular \pkg{sf} polygon.
}
