% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{get_srm4r}
\alias{get_srm4r}
\title{Get swissmapraster4r Environment data}
\usage{
get_srm4r(x)
}
\description{
Gets data stored in the package's own environment
}
\details{
Simply a wrapper around \code{get()} where \code{envir = } is set to the package's
default environment, \code{swissrastermapEnv}.
}
