% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_helper.R
\name{guess_scale}
\alias{guess_scale}
\title{Guess scale}
\usage{
guess_scale(extent, available_scales, factor = 1, outsize = NULL)
}
\arguments{
\item{extent}{A dataframe containing \code{xmin}, \code{xmax}, \code{ymin} and
\code{ymax} values}

\item{factor}{Usually, the scale is overestimated and smaller scales are desired.
This factor enables correction by dividing the calculated scale by the factor.
(e.g. a \code{factor = 2} returns a 1:25'000 scale when 1:50'000 has been calculated)}

\item{outsize}{An integer vector of the dimensions of the output window size in cm.
If left at \code{NULL} (default), the output window size is determined by
\code{dev.size("cm")}}

\item{available_cales}{An integer vector containing the available scales in the
format x in 1:1'000x (e.g. 25 in 1:25'000).}
}
\description{
Guess map scale based on the extent and the output window
}
\details{
This function takes an extent object (with \code{xmin}, \code{xmax},
\code{ymin} and \code{ymax} values)
as well as a set of available scales to determin the best suited scale.
The extent object (\code{xmin}, \code{xmax}, \code{ymin} and \code{ymax} values)
are expected in Meters while the available scales are expected to be the the
x in 1:1'000x (e.g. 25 in 1:25'000).
}
