% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdir_helper.R
\name{metainfo_from_filename}
\alias{metainfo_from_filename}
\title{Get Metadata from filename}
\usage{
metainfo_from_filename(filename, pattern)
}
\description{
Retrieves metadata from filename with the help of a predefined 'pattern'
}
\details{
Takes a filename (character string) and retrieves metadata from the filename
by consulting a predefined pattern. The whole process is somewhat similar to
specifying a format of a datetime object (see \code{strftime}). The pattern is
specified by placeholders (in the argument \code{pattern} ), which are in turn
specified in a dataframe (in the argument \code{search_pattern_dict}). Run
\code{get_srm4r("search_pattern_dict")} to see what these argument currently are.
}
