% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_colours.R
\name{raster_greyscale}
\alias{raster_greyscale}
\title{Threeband RGB to singleband GREYSCALE raster}
\usage{
raster_greyscale(raster_in, method = "ysrgb")
}
\arguments{
\item{raster_in}{An object of class RasterBrick.}

\item{method}{The method to use when turning RGB to greyscale.}
}
\description{
\code{rgb_raster2singleband} turns a threeband RGB Raster (\code{\link{brick}})
into a singleband GREYSCALE \code{\link{raster}} with a \code{\link{colortable}}
}
