% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{rgb_grey_weighted}
\alias{rgb_grey_weighted}
\title{Name}
\usage{
rgb_grey_weighted(raster_in, weights = c(0.3, 0.59, 0.11))
}
\arguments{
\item{param}{desc}
}
\description{
Short Desc
}
\details{
Long Desc
}
