% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_colours.R
\name{rgb_grey_ysrgb}
\alias{rgb_grey_ysrgb}
\title{Name}
\usage{
rgb_grey_ysrgb(raster_in, weights = c(0.2126, 0.7152, 0.0722))
}
\arguments{
\item{param}{desc}
}
\description{
Short Desc
}
\details{
Long Desc
}
