% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{rgb_raster2singleband}
\alias{rgb_raster2singleband}
\title{Threeband RGB to singleband raster}
\usage{
rgb_raster2singleband(rgb_raster)
}
\arguments{
\item{rgb_raster}{A character string pointing to a raster brick object or
an object of class RasterBrick. RBG values are assumed to range from 0-255}
}
\description{
\code{rgb_raster2singleband} turns a threeband RGB Raster (\code{\link{brick}})
into a singleband \code{\link{raster}} with a \code{\link{colortable}}
}
\details{
This function turns a three band RGB Raster Brick (from the Package \code{\link{raster}})
into a singleband raster with an included \code{\link{colortable}}. The function
requires the packages:  \code{\link{raster}}, \code{\link{rgdal}}
}
