% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{sample_points}
\alias{sample_points}
\title{Sample Points}
\format{An \code{sf} object containing point data
\describe{
  \item{NAME}{Name of the muncipality in its main language}
  \item{EINWOHNERZ}{Number of inhabitants}
  \item{geometry}{sfc_POINT data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
sample_points
}
\description{
Centroids of the 10 largest muncipalities ("Gemeinden") in Switzerland.
This dataset is based on \code{\link{sample_points}}
}
\keyword{datasets}
