% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_polygons}
\alias{sample_polygons}
\title{Sample Polygons}
\format{An \code{sf} object containing polygon data
\describe{
  \item{NAME}{Name of the muncipality in its main language}
  \item{EINWOHNERZ}{Number of inhabitants}
  \item{geometry}{sfc_POLYGON  data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
sample_polygons
}
\description{
Polygons of the 10 largest muncipalities ("Gemeinden") in Switzerland.
}
\keyword{datasets}
