% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdir_helper.R
\name{search_pattern}
\alias{search_pattern}
\title{Prints the search pattern nomenclature}
\usage{
search_pattern()
}
\description{
Prints the search pattern nomenclature
}
\details{
Retrieves the \code{search_pattern_dict} which is used in the
function \code{metainfo_from_filename}. Any characters not specified in
\code{search_pattern_dict} will be ignored in the '.pattern' file.
}
