% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdir_helper.R
\name{sf_remove_geom}
\alias{sf_remove_geom}
\title{Remove geometry from sf objects}
\usage{
sf_remove_geom(sf)
}
\arguments{
\item{sf}{Anything.}
}
\value{
The input, if it was \emph{not} of class \code{sf}. A \code{data.frame},
if it \emph{was}
}
\description{
Funtion to remove geometry, if the object \emph{is} in fact an sf object. Simply
Returns the input object if it isn't. This funtion is only necessary, because
\code{st_set_geometry(x,NULL)} returns an error if the input is not sf, and
\code{data.frame(x)} keep the geometry as a column.
}
