% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{show_extents}
\alias{show_extents}
\title{Show extents of Available Raste Files}
\usage{
show_extents(filedirectory, method = "ggplot2", interactive = F)
}
\arguments{
\item{filedirectory}{The file directory aquired by \code{init_fdir()}}

\item{method}{The method with which to visualize the data. "ggplot2" or "tmap"}

\item{interactive}{Should the plot be interactive or not? If TRUE, plotly()
is used when method is ggplot2.}
}
\description{
This function visualizes the extents of all available rasters. The plots are
faceted by scale and the area is colorized by resolution. The extents have an
alpha value so that overlapping extents can be detected.
}
\details{
This function requires ggplot2, and tmap and plotly for interactive plots
}
