% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{switzerland}
\alias{switzerland}
\title{switzerland}
\format{An \code{sf} object containing polygon data with four features.
\describe{
  \item{NAME}{Name of the country}
  \item{geometry}{sfc_POLYGON  data in EPSG 2056}
}}
\source{
\url{https://shop.swisstopo.admin.ch/de/products/landscape/boundaries3D}
}
\usage{
switzerland
}
\description{
Swissboundries3D Country data .
}
\keyword{datasets}
