% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{ysrgb}
\alias{ysrgb}
\title{Y_linear to Y_srgb}
\usage{
ysrgb(val)
}
\arguments{
\item{val}{The Y_linear Value to be turned into Y_srgb}
}
\description{
Helperfunction to turn Y_linear values to Y_srgb as described in
\link{https://en.wikipedia.org/wiki/Grayscale}
}
