% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdir.R
\name{fdir_init}
\alias{fdir_init}
\title{Initialize a "File Directory" by scanning all raster files}
\usage{
fdir_init(rootdir, add_geometry = T, filter = c("PK", "SMR", "LK",
  "TA"))
}
\arguments{
\item{rootdir}{Character string specifying the directory where the folders are stored}

\item{add_geometry}{Should the bounding box of each file be added as a geometry to \code{fdir}?}

\item{filter}{Names of the folders to look for. Only folders containing the character strings
specified here are included in the search.}
}
\description{
This function creates a "file directory" by scanning all available raster files
and extracting relevant metadata.
}
\details{
This function scans all folders in the root dirctory corresponding to the folders given by
\code{folders = }. It saves this directory in the package environment, and and
can later use this info to determin which rasters to import when running the
command \code{get_raster()}.

Currently, only files ending with "tif" are checked for extent, number of layers
and resolution and other informatin.

Parent folders (subfolders of \code{rootdir =}) are scanned non recursively for
raster data with the extention ".tif". These folders must adhere a specific structure:
\code{TYPE_Scale_EPSG_index}
\describe{
  \item{\code{TYPE}}{corresponds to the maptype. IN Switzerland this is typically \code{SMR}, \code{PK}, \code{LK} or \code{TA}. \strong{Must} be followed by a \code{_}}
  \item{\code{Scale}}{defines the map scale as x in 1:1'000x. \strong{Must} be followed by a \code{_}}
  \item{\code{EPSG}}{specifies the CRS of the containing raster data \strong{Must} be followed by a \code{_} \strong{only} if \code{name} is specified}
  \item{\code{index}}{Optional}
}
}
