% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{get_extent}
\alias{get_extent}
\title{Get extent of \code{sf} object}
\usage{
get_extent(features, x_add = 0, y_add = 0, method = "centroid",
  per_feature = T)
}
\arguments{
\item{param}{desc}
}
\description{
Get extent of \code{sf} object
}
\details{
Long Desc
}
