% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_functions.R
\name{get_extent}
\alias{get_extent}
\title{Name}
\usage{
get_extent(features, x_add, y_add, method = "centroid",
  per_feature = T)
}
\arguments{
\item{param}{desc}
}
\description{
Short Desc
}
\details{
Long Desc
}
