% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raster_helper.R
\name{get_srm4r}
\alias{get_srm4r}
\title{Get swissmapraster4r Environment data}
\usage{
get_srm4r(x)
}
\arguments{
\item{A}{character string specifying the object to retrieve from the environment.}
}
\description{
Gets data stored in the package's own environment
}
