% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{show_extents}
\alias{show_extents}
\title{Show extents of Available Raste Files}
\usage{
show_extents(method = "ggplot2", fdir = NULL)
}
\arguments{
\item{method}{The method with which to visualize the data: \code{ggplot2} or \code{tmap}
Wrap the output in \code{ggplotly()} to get an interactive ggplot output or set \code{tmap_mode}
to "view" go get an interactive tmap.}

\item{filedirectory}{The file directory aquired by \code{\link{fdir_init()}}. If left as
\code{NULL}, the \code{fdir} from the package Environment is taken.}
}
\description{
This function visualizes the extents of all available rasters. The plots are
faceted by scale and the area is colorized by resolution. The extents have an
alpha value so that overlapping extents can be detected.
}
\details{
This function requires ggplot2 and / or tmap, depending on the output type defined
in "method".
}
